Add files via upload
Because the dataset I used in "Calculating E(t)" didn't have the values for the parameters R and a, I estimated them using the least squares approach. I then used the calculated values in "Calculating E(t)". Since I'm not yet entirely sure of the data I will be using (and thus whether I'll have to do regression analysis), this is a side-thing for now.